+++ src/features/search-review-keyword/ui/ReviewWithPagination.tsx
@@ -1,0 +1,29 @@
+'use client';
+
+import {useParams, useSearchParams} from 'next/navigation';
+import {ReviewArticle, useSearchReviewsWithKeyword} from '@/entities/reviews';
+import Pagination from '@/widgets/pagination';
+import {SortKey} from '@/features/review-sorting/model/type';
+
+type Props = {
+  sort: SortKey;
+};
+
+export default function ReviewWithPagination({sort}: Props) {
+  const {keyword} = useParams<{keyword: string}>();
+  const searchParams = useSearchParams();
+
+  const currentPage = Number(searchParams.get('page')) || 1;
+  const {results, total_pages} = useSearchReviewsWithKeyword(keyword, currentPage, sort);
+
+  return (
+    <ul className="flex flex-col mb-6">
+      {results.map(searchReview => (
+        <li key={searchReview.board_id}>
+          <ReviewArticle searchReview={searchReview} />
+        </li>
+      ))}
+      <Pagination currentPage={currentPage} totalPage={total_pages} query={keyword} sort={sort} />
+    </ul>
+  );
+}
